% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath$get_file_paths}
\alias{PortrPath$get_file_paths}
\alias{PortrPath-class-get_file_paths}
\title{Get a named list of file paths}
\value{
A named list of file paths
}
\description{
For files specified in the shared config, get a named list of their local
paths
}
