% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath$root}
\alias{PortrPath$root}
\title{Get or set the current root}
\arguments{
\item{value}{the value of the root to be used. Either an absolute path
or ".PROJECT_ROOT"}
}
\value{
the current root, or nothing if setting
}
\description{
Sets the current root if assigned a value, or gets it if not
}
\seealso{
Other PortrPath: \code{\link{PortrPath$add_profile}},
  \code{\link{PortrPath$files}},
  \code{\link{PortrPath$get_file_paths}},
  \code{\link{PortrPath$new}},
  \code{\link{PortrPath$profile}}, \code{\link{PortrPath}}

Other path_access: \code{\link{PortrPath$files}},
  \code{\link{PortrPath$get_file_paths}}
}
\concept{PortrPath}
\concept{path_access}
