% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\docType{data}
\name{PortrPath-class}
\alias{PortrPath-class}
\alias{PortrPath}
\title{Main class for portrpaths}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PortrPath
}
\description{
R6 Reference class encapsulating all portrpaths behaviour and data
The current implementation will provide access to, essentially, a flat
Directory full of datafiles.
Local configuration, as well as providing an absolute root reference,
MIGHT also allow overriding of global options (To be decided)
}
\keyword{datasets}
