% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath}
\alias{PortrPath}
\title{Main class for portrpaths}
\description{
R6 Reference class encapsulating all portrpaths behaviour and data
}
\details{
The current implementation will provide access to, essentially, a flat
Directory full of datafiles.

Local configuration, as well as providing an absolute root reference,
MIGHT also allow overriding of global options (To be decided)
}
\seealso{
Other PortrPath: \code{\link{PortrPath$add_profile}},
  \code{\link{PortrPath$files}},
  \code{\link{PortrPath$get_file_paths}},
  \code{\link{PortrPath$new}},
  \code{\link{PortrPath$profile}},
  \code{\link{PortrPath$root}}
}
\concept{PortrPath}
\concept{class}
