% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths.R
\name{portrpaths-package}
\alias{portrpaths-package}
\title{Portable Paths for Network Drive Dependencies}
\description{
This tool allows configuration of paths on network drives to be machine
independant, and yaml configurable
}
