% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath$add_profile}
\alias{PortrPath$add_profile}
\title{Add a profile to the list of profiles}
\arguments{
\item{name}{the friendly name to refer to the profile}

\item{path}{the path the profile should refer to}
}
\value{
None
}
\description{
Profiles allow the user to easily switch between some favourite paths
}
\seealso{
Other PortrPath: \code{\link{PortrPath$get_file_paths}},
  \code{\link{PortrPath$new}},
  \code{\link{PortrPath$profile}},
  \code{\link{PortrPath$root}}, \code{\link{PortrPath}}
}
\concept{PortrPath}
