% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath$files}
\alias{PortrPath$files}
\title{Get a named list of file paths}
\arguments{
\item{value}{unused}
}
\value{
a list of files belonging to the object
}
\description{
For files specified in the shared config, get a named list of their local
paths
}
\seealso{
Other PortrPath: \code{\link{PortrPath$add_profile}},
  \code{\link{PortrPath$get_file_paths}},
  \code{\link{PortrPath$new}},
  \code{\link{PortrPath$profile}},
  \code{\link{PortrPath$root}}, \code{\link{PortrPath}}

Other path_access: \code{\link{PortrPath$get_file_paths}},
  \code{\link{PortrPath$root}}
}
\concept{PortrPath}
\concept{path_access}
