% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath$get_file_paths}
\alias{PortrPath$get_file_paths}
\title{Get a named list of file paths (Deprecated)}
\value{
A named list of file paths
}
\description{
This is deprecated, see \code{PortrPath$files}
}
\section{DEPRECATED}{

\code{PortrPath$get_file_Paths()} is the deprecated interface for this.
}

\seealso{
\link{PortrPath$files} for the \strong{recommended} accessor for file paths

Other PortrPath: \code{\link{PortrPath$add_profile}},
  \code{\link{PortrPath$files}},
  \code{\link{PortrPath$new}},
  \code{\link{PortrPath$profile}},
  \code{\link{PortrPath$root}}, \code{\link{PortrPath}}

Other path_access: \code{\link{PortrPath$files}},
  \code{\link{PortrPath$root}}
}
\concept{PortrPath}
\concept{path_access}
