% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath$new}
\alias{PortrPath$new}
\title{Inititalize A PortrPath object}
\arguments{
\item{local_config_path}{The location of the configuration file that define
local configuration parameters}

\item{shared_config_path}{The location of the configuration file shared
across the project}
}
\description{
Initializes new PortrPath object from configuration files given as arguments
}
\section{TODO}{

Document the input format
}

\seealso{
Other PortrPath: \code{\link{PortrPath$add_profile}},
  \code{\link{PortrPath$get_file_paths}},
  \code{\link{PortrPath$profile}},
  \code{\link{PortrPath$root}}, \code{\link{PortrPath}}
}
\concept{PortrPath}
