% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath$profile}
\alias{PortrPath$profile}
\title{Set the active profile}
\arguments{
\item{value}{the identifying string for the profile}
}
\value{
current profile if not being set
}
\description{
For set the current profile according to its id
Also write the d_root to file
}
\seealso{
Other PortrPath: \code{\link{PortrPath$add_profile}},
  \code{\link{PortrPath$get_file_paths}},
  \code{\link{PortrPath$new}},
  \code{\link{PortrPath$root}}, \code{\link{PortrPath}}
}
\concept{PortrPath}
