% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrpaths-class.R
\name{PortrPath-class}
\alias{PortrPath-class}
\alias{PortrPath}
\title{Main class for portrpaths}
\description{
R6 Reference class encapsulating all portrpaths behaviour and data
}
\details{
The current implementation will provide access to, essentially, a flat
Directory full of datafiles.

Local configuration, as well as providing an absolute root reference,
MIGHT also allow overriding of global options (To be decided)
}
